Change: Move calc_reward to evaluation module
